add icons to graph nodes
